remoting: do not log invalid raw data
Extracted from
https://crrev.com/c/3423288/9..15/remoting/host/security_key/security_key_message_handler.cc#b83

This causes UnicodeDecodeError in python script handling this log
e.g. asan_symbolize.py

Bug: 1288169
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3428532
Auto-Submit: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Joe Downing <joedow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#966238}

+++ remoting/host/security_key/security_key_message_handler.cc
@@ -79,8 +79,9 @@ void SecurityKeyMessageHandler::ProcessSecurityKeyMessage(
   } else if (message_type == SecurityKeyMessageType::REQUEST) {
     HandleSecurityKeyRequest(message->payload());
   } else {
+    // uint8_t is handled as char, so we use uint16_t to show number here.
     LOG(ERROR) << "Unknown message type: "
-               << static_cast<uint8_t>(message_type);
+               << static_cast<uint16_t>(message_type);
     SendMessage(SecurityKeyMessageType::UNKNOWN_COMMAND);
   }
 }
